+++ lib/marketInfo.ts
@@ -89,6 +89,26 @@ All other tokens settle at their defined minimum value of $0.1.`,
     isOpen: true,
   },
 
+  ISL: {
+    title: 'Indonesia Super League',
+    content: `The Indonesia Super League Performance Index is a close-ended digital market that reflects the current price of the performance tokens offered. This index operates under the principles of Haqq Mali to ensure an ethical and transparent trading environment.
+
+Initial Price Setting: Before the market opens, ShareMatch creates the initial price for each token based on historical oracle-based data from official Indonesia Super League statistics.
+
+Token Status: While the index itself is a series of close-ended markets, the underlying digital tokens are perpetual assets representing club performance.
+
+Secondary Market Trading: Once the index is open, a secondary market becomes available, allowing users to trade their tokens on the platform. This secondary market activity does not alter the original terms of the smart contract.
+
+Transparent Pricing: The pricing in the secondary market is solely determined by user supply and demand, ensuring a free and honest market with no interference or third-party data influence from the issuer (ShareMatch). Liquidity within the markets is provided by ShareMatch.
+
+Market Closure & Settlement: Trading is permitted until the clearly displayed closure of the market. At this point, the smart contract settles all positions with a defined value:
+
+The with profits token (i.e., the token at the top of the index) settles at its defined maximum value of $100.0.
+All other tokens settle at their defined minimum value of $0.1.`,
+    seasonDates: '1 July 2025 - 30 May 2026',
+    isOpen: true,
+  },
+
   WC: {
     title: 'FIFA World Cup Performance Index',
     content: `The FIFA World Cup Performance Index is a close-ended digital market that reflects the current price of the performance tokens offered. This index operates under the principles of Haqq Mali to ensure an ethical and transparent trading environment.